+++ CHANGELOG.md
@@ -1,5 +1,9 @@
 # Changelog
 
+#### 6th March July 2024 - Version 1.0.3
+
+-   Fix: Decrease minimum required PHP version to 7.4
+
 #### 15th July 2023 - Version 1.0.2
 
 -   Tweak: Added template system